Renamed model folders

+++ yelp-app/yelp-backend/src/main/java/yelp/dataset/oswego/yelpbackend/data_structure/weighted_graph/WeightedEdge.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class WeightedEdge implements Comparable<WeightedEdge> {
-    private int sourceID, destinationID;
+    private long sourceID, destinationID;
     private double weight;
     public int compareTo(WeightedEdge other) {return Double.compare(this.weight, other.weight);}
 }
